API-CALENDAR/requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
requests
pytz
python-dateutil